fix: guard ACP child callbacks
